TASK-1028 enforce identity authentication
By default ssh will try to connect to an authentication server if
SSH_AUTH_SOCK is define in environment variable. Since cron doesn't
define the varible, ssh is using a different mean to authenticate to
minarca server.

To ensure cron and UI are using the same method to authenticate, make
use of `-oIdentitiesOnly=yes`.